refactor(cli): use got and fix lint errors
